Connect commands with each other
