+++ cli/src/commands/schema/edit.ts
@@ -1,0 +1,15 @@
+import { BaseCommand } from '../../base.js';
+
+export default class EditSchema extends BaseCommand {
+  static description = 'Edit the schema of the current database';
+
+  static examples = [];
+
+  static flags = {};
+
+  static args = [];
+
+  async run(): Promise<void> {
+    // TODO
+  }
+}
